requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
pydantic
python-dotenv
celery
redis
